2004-02-13 16:25 UTC+0700 Andi Jahja <xharbour@cbn.net.id>   * source/vm/proc.c     ! casting.
  * source/rtl/gtgraph.c
  * source/rtl/hbsrlraw.c
  * source/rtl/gtwin/gtwin.c
  * source/rtl/gtwvt/gtwvt.c
  * source/rtl/gtalleg/gtalleg.c
    ! code clean-up ( mostly, consistent declaration of variable type to be
      compared with passed parameter)

    /*
    I am sorry but the little thing is like the following sample.
    It is known that every C compiler does automatically expand values upon
    comparing but IMHO, an explicit type declaration is better (CMIIW).

    void foo ( ULONG ulSize )
    {
       int ui; // int changed to ULONG
       ......

       if ( ulSize > ui )
       .......
    }
    */

+++ xharbour/source/rtl/gtalleg/gtalleg.c
@@ -1,5 +1,5 @@
 /*
- * $Id: gtalleg.c,v 1.19 2004/02/06 03:36:50 maurifull Exp $
+ * $Id: gtalleg.c,v 1.20 2004/02/06 17:07:28 jonnymind Exp $
  */
 
 /*
@@ -81,7 +81,7 @@ BOOL lClearInit = TRUE;
 static void hb_gt_DoCursor( void );
 
 static char *s_clipboard = NULL;
-static int s_clipsize = 0;
+static ULONG s_clipsize = 0;
 
 // I'm not sure of removing these (yet)
 // (they used to be static vars to center gt in hw screen, but now the
